% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualis.R
\name{get_qualis}
\alias{get_qualis}
\title{Get the 2019 Qualis CAPES database}
\usage{
get_qualis(file, ...)
}
\arguments{
\item{file}{Path to save the base (must be a CSV file)}

\item{...}{Other arguments passed by \code{\link[readr:write_delim]{readr::write_csv()}}}
}
\value{
Invisible, the updated base
}
\description{
Read the dataset from preliminary 2019 Qualis CAPES database, saving in a
CSV file according to the argument \code{file}. Data available in:
\url{https://github.com/enoches/Qualis_2019_preliminar}.
}
