% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qualiscapes}
\alias{qualiscapes}
\title{2019 Qualis CAPES database}
\format{
A data frame with 22,046 rows and 3 variables:
\describe{
\item{ISSN_2019}{The issn of the Journal/Congress}
\item{TITULO_2019}{The name of the Journal/Congress}
\item{ESTRATO_2019}{The Qualis CAPES of the Journal/Congress}
}
}
\source{
\url{https://github.com/enoches/Qualis_2019_preliminar}
}
\usage{
qualiscapes
}
\description{
A database with the preliminary 2019 Qualis CAPES.
}
\keyword{datasets}
